requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
matplotlib
jupyter
pyarrow
catboost
mysql-connector-python
python-dotenv
pefile

fastapi
uvicorn
python-multipart
jinja2
reportlab
